mess with structs to improve alignment

+++ config.h
@@ -55,7 +55,7 @@ static const Rule rules[] = {
 	{ "OMEdit",   NULL,       NULL,         1 << 3,     1,           -1,          false,           false,},
 	{ NULL,       NULL,       " - mpv",     1 << 4,     1,           -1,          false,           false,},
 	{ NULL,       NULL,       "ncmpcpp",    1 << 5,     1,           -1,          false,           false,},
-	{ NULL,       NULL,       "csv_plotter.py",~0,      0,           -1,          true,            false,},
+	{ NULL,       NULL,       "csv_plotter.py",(uint)~0,0,           -1,          true,            false,},
 	{ NULL,       NULL,       "clip.sh",    0,          0,           -1,          true,            false,},
 	{ NULL,       NULL,       "clip1.sh",   0,          0,           -1,          true,            false,},
 	{ NULL,       NULL,       "arqs.zsh",   0,          0,           -1,          true,            false,},
